+++ src/main/java/zos/shell/command/impl/ListCommand.java
@@ -7,7 +7,7 @@ import org.slf4j.LoggerFactory;
 import zos.shell.command.AbstractCommand;
 import zos.shell.command.CommandContext;
 import zos.shell.constants.Constants;
-import zos.shell.controller.container.ControllerFactories;
+import zos.shell.controller.factory.ControllerFactories;
 import zos.shell.response.ResponseStatus;
 import zos.shell.utility.DsnUtil;
