+++ src/main/java/zos/shell/controller/factory/AbstractDependencyControllerFactory.java
@@ -1,4 +1,4 @@
-package zos.shell.controller.container;
+package zos.shell.controller.factory;
 
 import zos.shell.controller.dependency.AbstractDependencyController;
 import zos.shell.controller.dependency.Dependency;
