+++ src/main/java/zos/shell/controller/factory/impl/ChangeControllerFactory.java
@@ -1,4 +1,4 @@
-package zos.shell.controller.container.impl;
+package zos.shell.controller.factory.impl;
 
 import org.beryx.textio.TextTerminal;
 import org.slf4j.Logger;
@@ -6,8 +6,8 @@ import org.slf4j.LoggerFactory;
 import zos.shell.controller.ChangeConnController;
 import zos.shell.controller.ChangeDirController;
 import zos.shell.controller.ChangeWinController;
-import zos.shell.controller.container.AbstractController;
-import zos.shell.controller.container.type.ChangeControllerType;
+import zos.shell.controller.factory.AbstractController;
+import zos.shell.controller.factory.type.ChangeControllerType;
 import zos.shell.service.change.ChangeConnService;
 import zos.shell.service.change.ChangeDirService;
 import zos.shell.service.change.ChangeWinService;
