+++ src/main/java/zos/shell/controller/factory/impl/DatasetControllerFactory.java
@@ -1,12 +1,12 @@
-package zos.shell.controller.container.impl;
+package zos.shell.controller.factory.impl;
 
 import org.beryx.textio.TextTerminal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import zos.shell.controller.*;
-import zos.shell.controller.container.AbstractDependencyControllerFactory;
-import zos.shell.controller.container.type.DatasetControllerType;
 import zos.shell.controller.dependency.Dependency;
+import zos.shell.controller.factory.AbstractDependencyControllerFactory;
+import zos.shell.controller.factory.type.DatasetControllerType;
 import zos.shell.service.checksum.CheckSumService;
 import zos.shell.service.dsn.concat.ConcatService;
 import zos.shell.service.dsn.copy.CopyService;
