+++ src/main/java/zos/shell/controller/factory/impl/GlobalControllerFactory.java
@@ -1,9 +1,9 @@
-package zos.shell.controller.container.impl;
+package zos.shell.controller.factory.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import zos.shell.controller.container.AbstractController;
-import zos.shell.controller.container.type.GlobalControllerType;
+import zos.shell.controller.factory.AbstractController;
+import zos.shell.controller.factory.type.GlobalControllerType;
 
 public class GlobalControllerFactory extends AbstractController<GlobalControllerType.Name> {
