+++ src/main/java/zos/shell/controller/factory/type/GlobalControllerType.java
@@ -1,4 +1,4 @@
-package zos.shell.controller.container.type;
+package zos.shell.controller.factory.type;
 
 public class GlobalControllerType {
